+++ scripts/test.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
-t3ize_sh="${0%/*}/../lib/t3ize.sh"
-source "$t3ize_sh"
+
+source $(t3x -T)
 
 info "information message"
 debug "debug message"