+++ tools/pi/scripts/disable-ufw.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/bash
+source $(t3x -T)
+
+echo "disabling ufw"
+sudo ufw disable